+++ pyccel/codegen/printing/ccode.py
@@ -46,7 +46,6 @@ from pyccel.ast.numpytypes import NumpyInt8Type, NumpyInt16Type, NumpyInt32Type,
 from pyccel.ast.numpytypes import NumpyFloat32Type, NumpyFloat64Type, NumpyComplex64Type, NumpyComplex128Type
 from pyccel.ast.numpytypes import NumpyNDArrayType, numpy_precision_map
 from pyccel.ast.cudatypes  import CudaArrayType
-from pyccel.ast.cudaext    import CudaFull
 
 from pyccel.ast.type_annotations import VariableTypeAnnotation
 
@@ -61,6 +60,7 @@ from pyccel.ast.variable import InhomogeneousTupleVariable
 from pyccel.ast.c_concepts import ObjectAddress, CMacro, CStringExpression, PointerCast, CNativeInt
 
 from pyccel.codegen.printing.codeprinter import CodePrinter
+
 
 
 from pyccel.errors.errors   import Errors
@@ -243,6 +243,7 @@ c_imports = {n : Import(n, Module(n, (), ())) for n in
                  'stdbool',
                  'assert',
                  'numpy_c']}
+
 
 import_header_guard_prefix = {'Set_extensions' : '_TOOLS_SET'}
 
@@ -1319,6 +1320,7 @@ class CCodePrinter(CodePrinter):
         >>> self.get_declare_type(v)
         't_ndarray*'
         """
+        from pyccel.codegen.printing.cucode import cu_imports
         class_type = expr.class_type
         rank  = expr.rank
 
@@ -1326,10 +1328,14 @@ class CCodePrinter(CodePrinter):
             if isinstance(expr.class_type, (HomogeneousSetType, HomogeneousListType)):
                 dtype = self.get_c_type(expr.class_type)
                 return dtype
-            if isinstance(expr.class_type,(HomogeneousTupleType, NumpyNDArrayType, CudaArrayType)):
+            if isinstance(expr.class_type,(HomogeneousTupleType, NumpyNDArrayType)):
                 if expr.rank > 15:
                     errors.report(UNSUPPORTED_ARRAY_RANK, symbol=expr, severity='fatal')
                 self.add_import(c_imports['ndarrays'])
+                dtype = 't_ndarray'
+            elif isinstance(expr.class_type, CudaArrayType):
+                self.add_import(c_imports['ndarrays'])
+                self.add_import(cu_imports['cuda_ndarrays'])
                 dtype = 't_ndarray'
             else:
                 errors.report(PYCCEL_RESTRICTION_TODO+' (rank>0)', symbol=expr, severity='fatal')