+++ pyccel/codegen/utilities.py
@@ -37,13 +37,15 @@ external_libs = {"stc"  : "STC/include",}
 # The compile object folder will be in the pyccel dirpath
 internal_libs = {
     "ndarrays"     : ("ndarrays", CompileObj("ndarrays.c",folder="ndarrays")),
-    "cuda_ndarrays": ("cuda_ndarrays", CompileObj("cuda_ndarrays.cu",folder="ndarrays")),
+    "cuda_ndarrays": ("cuda_ndarrays", CompileObj("cuda_ndarrays.cu",folder="cuda_ndarrays")),
     "pyc_math_f90" : ("math", CompileObj("pyc_math_f90.f90",folder="math")),
     "pyc_math_c"   : ("math", CompileObj("pyc_math_c.c",folder="math")),
-    "cwrapper"     : ("cwrapper", CompileObj("cwrapper.c",folder="cwrapper", accelerators=('python',))),
+    "cwrapper"     : ("cwrapper", CompileObj("cwrapper.c",folder="cwrapper",
+                                             accelerators=('python',))),
     "numpy_f90"    : ("numpy", CompileObj("numpy_f90.f90",folder="numpy")),
     "numpy_c"      : ("numpy", CompileObj("numpy_c.c",folder="numpy")),
-    "STC_Extensions" : ("STC_Extensions", CompileObj("Set_Extensions.h",folder="STC_Extensions", has_target_file = False)),
+    "Set_extensions" : ("STC_Extensions", CompileObj("Set_Extensions.h",
+                                             folder="STC_Extensions", has_target_file = False)),
 }
 internal_libs["cwrapper_ndarrays"] = ("cwrapper_ndarrays", CompileObj("cwrapper_ndarrays.c",folder="cwrapper_ndarrays",
                                                              accelerators = ('python',),
@@ -144,6 +146,9 @@ def copy_internal_library(lib_folder, pyccel_dirpath, extra_files = None):
         if to_create:
             # Copy all files from the source to the destination
             shutil.copytree(lib_path, lib_dest_path)
+            dst_files = [os.path.relpath(os.path.join(root, f), lib_dest_path) \
+                    for root, dirs, files in os.walk(lib_dest_path) \
+                    for f in files if not f.endswith('.lock')]
             # Create any requested extra files
             if extra_files:
                 for filename, contents in extra_files.items():