+++ pyccel/cuda/__init__.py
@@ -9,8 +9,4 @@ from .cuda_sync_primitives    import synchronize
 from .cuda_arrays             import host_empty
 from .cuda_arrays             import device_empty
 
-from .cuda_thread_indexing               import threadIdx
-from .cuda_thread_indexing               import blockIdx
-from .cuda_thread_indexing               import blockDim
-
-__all__ = ['synchronize', 'host_empty', 'device_empty', 'threadIdx', 'blockIdx', 'blockDim']
+__all__ = ['synchronize', 'host_empty', 'device_empty']