+++ pyccel/stdlib/cuda_ndarrays/cuda_ndarrays.h
@@ -2,15 +2,14 @@
 # define CUDA_NDARRAYS_H
 
 # include <cuda_runtime.h>
-# include <iostream>
 #include "../ndarrays/ndarrays.h"
 
 
 
 enum e_memory_locations
 {
-        allocateMemoryOnHost,
-        allocateMemoryOnDevice
+        HostMemory,
+        DeviceMemory
 };
 
 
@@ -20,8 +19,5 @@ enum e_memory_locations location);
 int32_t cuda_free_host(t_ndarray  arr);
 __host__ __device__
 int32_t cuda_free(t_ndarray  arr);
-
-
-using namespace std;
 
 #endif