+++ CHANGELOG.md
@@ -12,6 +12,7 @@ All notable changes to this project will be documented in this file.
 -   #42 : Add Cuda module to Pyccel. Add support for `cuda.synchronize` function.
 -   #41 : Add support for custom device in`cuda`.
 -   #69 : Improve kernel decorator.
+-   #64 : Add support for `cuda.device_empty` function.
 
 ## \[UNRELEASED\]